requirements.txt
Django==2.0
django-extensions==2.1.6
psycopg2-binary==2.7.7
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
gunicorn==19.9.0
json-logging-py==0.2
